chore(js): fixed the issue with double notifications request triggered
