+++ packages/js/src/ui/helpers/createInfiniteScroll.ts
@@ -50,6 +50,7 @@ export function createInfiniteScroll<T>(fetcher: (page: number) => Promise<{ dat
   });
 
   const reset = async () => {
+    setData([]);
     setInitialLoading(true);
     setEnd(false);
